refactor(forks/lstar): move multi-signature types out of crypto layer
The crypto subspec imported AggregationBits, Slot, ValidatorIndex, and
ValidatorIndices from the consensus layer to define TypeOneMultiSignature
and TypeTwoMultiSignature. That layering inversion forced three mid-file
deferred imports with # noqa: E402 in lstar/containers.py and made the
crypto layer reach into consensus for types it should never know about.

Move the multi-signature classes (plus AggregationError) into the
consensus layer, where they belong as domain-typed wrappers around the
crypto byte-level primitives. The crypto layer keeps only the Rust
prover bindings.

Slot moves into its own small module so the crypto layer can name a
slot without pulling the full consensus container module. The crypto
API still uses Slot, not Uint64 — the layering fix preserves semantics.

Net result:
- crypto/xmss/aggregation.py shrinks from 348 lines to 38; no consensus
  imports remain.
- crypto/xmss/{containers,interface}.py import Slot from lstar/slot.py.
- lstar/containers.py promotes all three previously deferred imports
  (HISTORICAL_ROOTS_LIMIT, multi-sig types, PublicKey/Signature) to the
  top of the file. The # noqa: E402 block is gone.
- 17 importers updated to fetch TypeOneMultiSignature etc. from the
  consensus layer.
