+++ packages/testing/src/consensus_testing/keys.py
@@ -43,7 +43,6 @@ from typing import ClassVar, Literal
 from lean_spec.config import LEAN_ENV
 from lean_spec.spec.crypto.koalabear import Fp
 from lean_spec.spec.crypto.merkleization import hash_tree_root
-from lean_spec.spec.crypto.xmss.aggregation import TypeOneMultiSignature
 from lean_spec.spec.crypto.xmss.constants import TARGET_CONFIG
 from lean_spec.spec.crypto.xmss.containers import (
     PublicKey,
@@ -63,7 +62,11 @@ from lean_spec.spec.crypto.xmss.types import (
     Randomness,
 )
 from lean_spec.spec.forks import Slot, ValidatorIndex
-from lean_spec.spec.forks.lstar.containers import AggregatedAttestations, AttestationData
+from lean_spec.spec.forks.lstar.containers import (
+    AggregatedAttestations,
+    AttestationData,
+    TypeOneMultiSignature,
+)
 from lean_spec.spec.ssz import Bytes32, Uint64
 
 KeyRole = Literal["attestation", "proposal"]
