+++ packages/testing/src/consensus_testing/test_types/aggregated_attestation_spec.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 from lean_spec.base import CamelModel
-from lean_spec.spec.crypto.xmss.aggregation import TypeOneMultiSignature
 from lean_spec.spec.forks import Checkpoint, Slot, ValidatorIndex, ValidatorIndices
 from lean_spec.spec.forks.lstar.containers import (
     AggregatedAttestation,
@@ -11,6 +10,7 @@ from lean_spec.spec.forks.lstar.containers import (
     AttestationData,
     Block,
     State,
+    TypeOneMultiSignature,
 )
 from lean_spec.spec.ssz import ByteList512KiB, Bytes32
