+++ packages/testing/src/consensus_testing/test_types/gossip_aggregated_attestation_spec.py
@@ -3,13 +3,13 @@
 from __future__ import annotations
 
 from lean_spec.base import CamelModel
-from lean_spec.spec.crypto.xmss.aggregation import TypeOneMultiSignature
 from lean_spec.spec.forks import Checkpoint, Slot, ValidatorIndex, ValidatorIndices
 from lean_spec.spec.forks.lstar.containers import (
     AttestationData,
     Block,
     SignedAggregatedAttestation,
     State,
+    TypeOneMultiSignature,
 )
 from lean_spec.spec.ssz import ByteList512KiB, Bytes32
