+++ src/lean_spec/spec/crypto/xmss/containers.py
@@ -5,7 +5,7 @@ from typing import override
 from pydantic import field_serializer, model_serializer
 
 from lean_spec.base import StrictBaseModel
-from lean_spec.spec.forks.lstar.containers import Slot
+from lean_spec.spec.forks.lstar.slot import Slot
 from lean_spec.spec.ssz import Uint64
 from lean_spec.spec.ssz.container import Container
