+++ tests/lean_spec/spec/forks/lstar/forkchoice/test_compute_block_weights.py
@@ -5,10 +5,9 @@ from __future__ import annotations
 import pytest
 
 from lean_spec.spec.crypto.merkleization import hash_tree_root
-from lean_spec.spec.crypto.xmss.aggregation import TypeOneMultiSignature
 from lean_spec.spec.forks import Checkpoint, Slot, ValidatorIndex, ValidatorIndices
 from lean_spec.spec.forks.lstar import Store
-from lean_spec.spec.forks.lstar.containers import AttestationData
+from lean_spec.spec.forks.lstar.containers import AttestationData, TypeOneMultiSignature
 from lean_spec.spec.forks.lstar.spec import LstarSpec
 from lean_spec.spec.ssz.byte_arrays import ByteList512KiB, Bytes32
 from tests.lean_spec.helpers import make_bytes32, make_signed_block
