+++ tests/lean_spec/spec/forks/lstar/forkchoice/test_store_pruning.py
@@ -1,8 +1,8 @@
 """Tests for Store attestation data pruning."""
 
-from lean_spec.spec.crypto.xmss.aggregation import TypeOneMultiSignature
 from lean_spec.spec.forks import Slot, ValidatorIndex, ValidatorIndices
 from lean_spec.spec.forks.lstar import AttestationSignatureEntry, Store
+from lean_spec.spec.forks.lstar.containers import TypeOneMultiSignature
 from lean_spec.spec.forks.lstar.spec import LstarSpec
 from lean_spec.spec.ssz import ByteList512KiB, Bytes32
 from tests.lean_spec.helpers import (
