+++ tests/lean_spec/spec/forks/lstar/forkchoice/test_validator.py
@@ -5,7 +5,6 @@ from consensus_testing.keys import XmssKeyManager
 
 from lean_spec.node.chain.clock import Interval
 from lean_spec.spec.crypto.merkleization import hash_tree_root
-from lean_spec.spec.crypto.xmss.aggregation import TypeOneMultiSignature
 from lean_spec.spec.forks import Checkpoint, Slot, ValidatorIndex
 from lean_spec.spec.forks.lstar import AttestationSignatureEntry, Store
 from lean_spec.spec.forks.lstar.containers import (
@@ -15,6 +14,7 @@ from lean_spec.spec.forks.lstar.containers import (
     BlockBody,
     Config,
     SignedAttestation,
+    TypeOneMultiSignature,
 )
 from lean_spec.spec.forks.lstar.spec import LstarSpec
 from lean_spec.spec.ssz import Bytes32, Uint64
